requirements.txt
requests
python-dotenv
# Dependências principais
psycopg2-binary
playwright
requests
apscheduler
anthropic
pytest
pytest-mock
pgvector
sqlalchemy
alembic
fastapi
uvicorn
pydantic
atlassian-python-api
mcp
qdrant-client
# Adicione aqui as dependências do seu projeto